Module CPath

type t = private string

val make : string list -> t

make path_components build a path from its components

val relative : t -> string -> t

relative path string build a path relative to an existing one

val choose_existing : t list -> t option

choose_existing paths will return Some f for the first file f in paths that exists, None otherwise.