Module Coqdeplib.Fl

val findlib_deep_resolve : file:string -> package:string -> string list * string list

findlib_deep_resolve ~file ~package tries to locate a .cmxs for a given package. It also searches for its dependencies.

Dependencies of coq-core.toplevel (i.e. coqc) are ignored.

file is used for error messages.

If a META file for package is found, it will try to use it to resolve the path to the .cmxs, and return a relative path to both. If not, it errors.