Module DeclareUniv

exception AlreadyDeclared of string option * Names.Id.t

Also used by Declare for constants, DeclareInd for inductives, etc. Containts object_kind , id.

val declare_univ_binders : Names.GlobRef.t -> UState.named_universes_entry -> unit

Internally used to declare names of universes from monomorphic constants/inductives. Noop on polymorphic references.

val name_mono_section_univs : Univ.Level.Set.t -> unit

Internally used to name universes associated with no particular constant in a section.

val do_universe : poly:bool -> Names.lident list -> unit

Command Universes.

val do_constraint : poly:bool -> Constrexpr.univ_constraint_expr list -> unit

Command Constraint.

val add_constraint_source : Names.GlobRef.t -> Univ.ContextSet.t -> unit

val constraint_sources : unit -> (Names.GlobRef.t * Univ.Constraints.t) list

Returns constraints associated to globrefs, newest first.