Module Dnet.Make

Parameters

module T : Datatype

module Ident : Stdlib.Set.OrderedType

module Meta : Stdlib.Set.OrderedType

Signature

type t

type ident = Ident.t

provided identifier type

type meta = Meta.t

provided metavariable type

type 'a structure = 'a T.t

provided parametrized datastructure

module Idset : Stdlib.Set.S with type elt = ident

returned sets of solutions

type term_pattern =

| Term of term_pattern structure

| Meta of meta

a pattern is a term where each node can be a unification variable

val empty : t

val add : t -> term_pattern -> ident -> t

add t w i adds a new association (w,i) in t.

val find_all : t -> Idset.t

find_all t returns all identifiers contained in t.

val find_match : term_pattern -> t -> Idset.t

find_match p t returns identifiers of all terms matching p in t.