Module EConstr.ESorts

type t

Type of sorts up-to universe unification. Essentially a wrapper around Sorts.t so that normalization is ensured statically.

val make : Sorts.t -> t

Turn a sort into an up-to sort.

val kind : Evd.evar_map -> t -> Sorts.t

Returns the view into the current sort. Note that the kind of a variable may change if the unification state of the evar map changes.