Module Eqschemes

This file builds schemes relative to equality inductive types

Builds a left-to-right rewriting scheme for an equality type

val rew_l2r_dep_scheme_kind : Ind_tables.individual Ind_tables.scheme_kind

val rew_l2r_scheme_kind : Ind_tables.individual Ind_tables.scheme_kind

val rew_r2l_forward_dep_scheme_kind : Ind_tables.individual Ind_tables.scheme_kind

val rew_l2r_forward_dep_scheme_kind : Ind_tables.individual Ind_tables.scheme_kind

val rew_r2l_dep_scheme_kind : Ind_tables.individual Ind_tables.scheme_kind

val rew_r2l_scheme_kind : Ind_tables.individual Ind_tables.scheme_kind

Builds a symmetry scheme for a symmetrical equality type

val sym_scheme_kind : Ind_tables.individual Ind_tables.scheme_kind

val sym_involutive_scheme_kind : Ind_tables.individual Ind_tables.scheme_kind

Builds a congruence scheme for an equality type

val congr_scheme_kind : Ind_tables.individual Ind_tables.scheme_kind