Module HConstr.Tbl

Imperative tables indexed by HConstr.t. The interfaces exposed are the same as Hashtbl but are not guaranteed to be implemented by Hashtbl.

type key = t

type 'a t

val find_opt : 'a t -> key -> 'a option

val add : 'a t -> key -> 'a -> unit

val create : int -> 'a t