Module Instr

Platform-specific Implementation of a global instruction counter.

val read_counter : unit -> (Stdlib.Int64.t, string) Stdlib.Result.t

read_counter () returns the current value of the instruction counter, or an error message indicating why it could not be obtained. Note that counts returned by this function are not meaningful in absolute terms. To measure how many instructions it takes to run a piece of code, read the counter at the start and at the end, and compute the difference.