Module Int.PMap

This is a (partial) implementation of a Map interface on integers, except that it internally uses persistent arrays. This ensures O(1) accesses in non-backtracking cases. It is thus better suited for zero-starting, contiguous keys, or otherwise a lot of space will be empty. To keep track of the present keys, a binary tree is also used, so that adding a key is still logarithmic. It is therefore essential that most of the operations are accesses and not add/removes.

type key = int

type 'a t

val empty : 'a t

val is_empty : 'a t -> bool

val mem : key -> 'a t -> bool

val add : key -> 'a -> 'a t -> 'a t

val singleton : key -> 'a -> 'a t

val remove : key -> 'a t -> 'a t

val iter : (key -> 'a -> unit) -> 'a t -> unit

val fold : (key -> 'a -> 'b -> 'b) -> 'a t -> 'b -> 'b

val for_all : (key -> 'a -> bool) -> 'a t -> bool

val exists : (key -> 'a -> bool) -> 'a t -> bool

val find : key -> 'a t -> 'a

val domain : 'a t -> Set.t

val cast : 'a t -> 'a Map.t