Module Polynomial.ProofFormat

Proof format used by the proof-generating procedures. It is fairly close to Coq format but a bit more liberal.

It is used for proofs over Z, Q, R. However, certain constructions e.g. CutPrf are only relevant for Z.

type prf_rule =

| Annot of string * prf_rule

| Hyp of int

| Def of int

| Ref of int

| Cst of NumCompat.Q.t

| Zero

| Square of Vect.t

| MulC of Vect.t * prf_rule

| Gcd of NumCompat.Z.t * prf_rule

| MulPrf of prf_rule * prf_rule

| AddPrf of prf_rule * prf_rule

| CutPrf of prf_rule

| LetPrf of prf_rule * prf_rule

type proof =

| Done

| Step of int * prf_rule * proof

| Split of int * Vect.t * proof * proof

| Enum of int * prf_rule * Vect.t * prf_rule * proof list

| ExProof of int * int * int * var * var * var * proof

val pr_size : prf_rule -> NumCompat.Q.t

val pr_rule_max_def : prf_rule -> int

val pr_rule_max_hyp : prf_rule -> int

val proof_max_def : proof -> int

val normalise_proof : int -> proof -> int * proof

val output_prf_rule : Stdlib.out_channel -> prf_rule -> unit

val output_proof : Stdlib.out_channel -> proof -> unit

val add_proof : prf_rule -> prf_rule -> prf_rule

val mul_cst_proof : NumCompat.Q.t -> prf_rule -> prf_rule

val mul_proof : prf_rule -> prf_rule -> prf_rule

module Env : sig ... end

val compile_proof : Env.t -> proof -> Micromega.zArithProof

val cmpl_prf_rule : ('a Micromega.pExpr -> 'a Micromega.pol) -> (NumCompat.Q.t -> 'a) -> Env.t -> prf_rule -> 'a Micromega.psatz

val proof_of_farkas : prf_rule Mutils.IMap.t -> Vect.t -> prf_rule

val simplify_proof : proof -> proof * Mutils.ISet.t