Module Names.MBId

type t

Unique names for bound modules. Each call to make constructs a fresh unique identifier.

val equal : t -> t -> bool

Equality over unique bound names.

val compare : t -> t -> int

Comparison over unique bound names.

val hash : t -> int

Hash over unique bound names.

val make : DirPath.t -> Id.t -> t

The first argument is a file name, to prevent conflict between different files.

val repr : t -> int * Id.t * DirPath.t

Reverse of make.

val to_id : t -> Id.t

Return the identifier contained in the argument.

val to_string : t -> string

Encode as a string (not to be used for user-facing messages).

val debug_to_string : t -> string

Same as to_string, but outputs extra information related to debug.