Module Proofview_monad.Info

Info trace.

type tag =

| Msg of lazy_msg	(*
A simple message
*)

| Tactic of lazy_msg	(*
A tactic call
*)

| Dispatch	(*
A call to tclDISPATCH/tclEXTEND
*)

| DBranch	(*
A special marker to delimit individual branch of a dispatch.
*)

The type of the tags for info.

type state = tag Trace.incr

type tree = tag Trace.forest

val print : Environ.env -> Evd.evar_map -> tree -> Pp.t

val collapse : int -> tree -> tree

collapse n t flattens the first n levels of Tactic in an info trace, effectively forgetting about the n top level of names (if there are fewer, the last name is kept).