Module Proofview_monad.InfoL

Lens and utilities pertaining to the info trace

val record_trace : 'a Logical.t -> 'a Logical.t

record_trace t behaves like t and compute its info trace.

val update : ( Info.state -> Info.state ) -> unit Logical.t

val opn : Info.tag -> unit Logical.t

val close : unit Logical.t

val leaf : Info.tag -> unit Logical.t

val tag : Info.tag -> 'a Logical.t -> 'a Logical.t

tag a t opens tag a runs t then closes the tag.