Module Relevanceops

We can take advantage of non-cumulativity of SProp to avoid fully retyping terms when we just want to know if they inhabit some proof-irrelevant type.

val relevance_of_term : Environ.env -> Constr.constr -> Sorts.relevance

val relevance_of_rel : Environ.env -> int -> Sorts.relevance

val relevance_of_var : Environ.env -> Names.Id.t -> Sorts.relevance

val relevance_of_constant : Environ.env -> Names.Constant.t -> Sorts.relevance

val relevance_of_constructor : Environ.env -> Names.constructor -> Sorts.relevance

val relevance_of_projection : Environ.env -> Names.Projection.t -> Sorts.relevance