Module Summary.Stage

This module registers the declaration of global tables, which will be kept in synchronization during the various backtracks of the system.

type t =

| Synterp

| Interp

We distinguish two stages and separate the system state accordingly. Synterp is the syntactic interpretation phase, i.e. vernacular parsing and execution of commands having an effect on parsing. Interp is the interpretation phase, where standard commands are executed.

val equal : t -> t -> bool