Module type Summary.FrozenStage

The type frozen is a snapshot of the states of all the registered tables of the system.

type frozen

val empty_frozen : frozen

val freeze_summaries : unit -> frozen

val make_marshallable : frozen -> frozen

val unfreeze_summaries : ?partial:bool -> frozen -> unit

val init_summaries : unit -> unit

val project_from_summary : frozen -> 'a Dyn.tag -> 'a