Module Termops.Internal

NOTE: to print terms you always want to use functions in Printer, not these ones which are for very special cases.

val debug_print_constr : Evd.evar_map -> EConstr.constr -> Pp.t

debug printers: print raw form for terms with evar-substitution.

val print_constr_env : Environ.env -> Evd.evar_map -> EConstr.constr -> Pp.t

Pretty-printer hook: print_constr_env env sigma c will pretty print c if the pretty printing layer has been linked into the Coq binary.

val set_print_constr : (Environ.env -> Evd.evar_map -> EConstr.constr -> Pp.t) -> unit

set_print_constr f sets f to be the pretty printer

val print_named_context : Environ.env -> Evd.evar_map -> Pp.t

Printers for contexts

val pr_rel_decl : Environ.env -> Evd.evar_map -> Constr.rel_declaration -> Pp.t

val print_rel_context : Environ.env -> Evd.evar_map -> Pp.t

val print_env : Environ.env -> Evd.evar_map -> Pp.t

val print_kconstr : Environ.env -> Evd.evar_map -> Evd.econstr -> Pp.t