Module Trie

Generic functorized trie data structure.

module type S = sig ... end

A trie is a generalization of the map data structure where the keys are themselves lists.

module type Grp = sig ... end

module Make (Label : Stdlib.Set.OrderedType) (Data : Grp) : S with type label = Label.t and type data = Data.t

Generating functor, for a given type of labels and data.