Module Level.Map

include CMap.ExtS with type key = t and module Set := Set

The underlying Map library

include CSig.MapS with type key = t

type key = t

type +'a t

val empty : 'a t

val is_empty : 'a t -> bool

val mem : key -> 'a t -> bool

val add : key -> 'a -> 'a t -> 'a t

val update : key -> ('a option -> 'a option) -> 'a t -> 'a t

val singleton : key -> 'a -> 'a t

val remove : key -> 'a t -> 'a t

val merge : (key -> 'a option -> 'b option -> 'c option) -> 'a t -> 'b t -> 'c t

val union : (key -> 'a -> 'a -> 'a option) -> 'a t -> 'a t -> 'a t

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

val equal : ('a -> 'a -> bool) -> 'a t -> 'a t -> bool

val iter : (key -> 'a -> unit) -> 'a t -> unit

val fold : (key -> 'a -> 'b -> 'b) -> 'a t -> 'b -> 'b

val for_all : (key -> 'a -> bool) -> 'a t -> bool

val exists : (key -> 'a -> bool) -> 'a t -> bool

val filter : (key -> 'a -> bool) -> 'a t -> 'a t

val partition : (key -> 'a -> bool) -> 'a t -> 'a t * 'a t

val cardinal : 'a t -> int

val bindings : 'a t -> (key * 'a) list

val min_binding : 'a t -> key * 'a

val max_binding : 'a t -> key * 'a

val choose : 'a t -> key * 'a

val choose_opt : 'a t -> (key * 'a) option

val split : key -> 'a t -> 'a t * 'a option * 'a t

val find : key -> 'a t -> 'a

val find_opt : key -> 'a t -> 'a option

val map : ('a -> 'b) -> 'a t -> 'b t

val mapi : (key -> 'a -> 'b) -> 'a t -> 'b t

val get : key -> 'a t -> 'a

Same as find but fails an assertion instead of raising Not_found

val set : key -> 'a -> 'a t -> 'a t

Same as add, but expects the key to be present, and thus faster.

raises Not_found
when the key is unbound in the map.

val modify : key -> (key -> 'a -> 'a) -> 'a t -> 'a t

Apply the given function to the binding of the given key.

raises Not_found
when the key is unbound in the map.

val domain : 'a t -> Set.t

Recover the set of keys defined in the map.

val bind : (key -> 'a) -> Set.t -> 'a t

bind f s transform the set x1; ...; xn into x1 := f x1; ...; xn := f xn.

val fold_left : (key -> 'a -> 'b -> 'b) -> 'a t -> 'b -> 'b

Alias for fold, to easily track where we depend on fold order.

val fold_right : (key -> 'a -> 'b -> 'b) -> 'a t -> 'b -> 'b

Folding keys in decreasing order.

val height : 'a t -> int

An indication of the logarithmic size of a map

val filter_range : (key -> int) -> 'a t -> 'a t

find_range in_range m Given a comparison function in_range x, that tests if x is below, above, or inside a given range filter_range returns the submap of m whose keys are in range. Note that in_range has to define a continouous range.

val filter_map : (key -> 'a -> 'b option) -> 'a t -> 'b t

Like map but keeping only bindings mapped to Some

val of_list : (key * 'a) list -> 'a t

Turns an association list into a map

val symmetric_diff_fold : (key -> 'a option -> 'a option -> 'b -> 'b) -> 'a t -> 'a t -> 'b -> 'b

symmetric_diff f ml mr acc will efficiently fold over the difference between ml and mr, assumed that they share most of their internal structure. A call to f k vl vr means that if vl is Some, then k exists in ml. Similarly, if vr is Some, then k exists in mr. If both vl and vr are Some, then vl != vr.

module Smart : sig ... end

module Monad (M : CMap.MonadS) : sig ... end

Fold operators parameterized by any monad.

val lunion : 'a t -> 'a t -> 'a t

lunion x y favors the bindings in the first map.

val diff : 'a t -> 'a t -> 'a t

diff x y removes bindings from x that appear in y (whatever the value).

val pr : (key -> Pp.t) -> ('a -> Pp.t) -> 'a t -> Pp.t

Pretty-printing