Module UnivGen

type univ_length_mismatch = {

actual : int;

expect : int;

}

exception UniverseLengthMismatch of univ_length_mismatch

Side-effecting functions creating new universe levels.

val new_univ_global : unit -> Univ.UGlobal.t

val fresh_level : unit -> Univ.Level.t

val new_global_univ : unit -> Univ.Universe.t Univ.in_universe_context_set

Build a fresh instance for a given context, its associated substitution and the instantiated constraints.

val fresh_instance : Univ.AbstractContext.t -> Univ.Instance.t Univ.in_universe_context_set

val fresh_instance_from : ?loc:Loc.t -> Univ.AbstractContext.t -> Univ.Instance.t option -> Univ.Instance.t Univ.in_universe_context_set

val fresh_sort_in_family : Sorts.family -> Sorts.t Univ.in_universe_context_set

val fresh_constant_instance : Environ.env -> Names.Constant.t -> Constr.pconstant Univ.in_universe_context_set

val fresh_inductive_instance : Environ.env -> Names.inductive -> Constr.pinductive Univ.in_universe_context_set

val fresh_constructor_instance : Environ.env -> Names.constructor -> Constr.pconstructor Univ.in_universe_context_set

val fresh_array_instance : Environ.env -> Univ.Instance.t Univ.in_universe_context_set

val fresh_global_instance : ?loc:Loc.t -> ?names:Univ.Instance.t -> Environ.env -> Names.GlobRef.t -> Constr.constr Univ.in_universe_context_set

val fresh_universe_context_set_instance : Univ.ContextSet.t -> Univ.universe_level_subst * Univ.ContextSet.t

Get fresh variables for the universe context. Useful to make tactics that manipulate constrs in universe contexts polymorphic.

val constr_of_monomorphic_global : Environ.env -> Names.GlobRef.t -> Constr.constr

Create a fresh global in the environment argument, without side effects. BEWARE: this raises an error on polymorphic constants/inductives: the constraints should be properly added to an evd. See Evd.fresh_global, Evarutil.new_global, and pf_constr_of_global for the proper way to get a fresh copy of a polymorphic global reference.