Module UnivProblem

Constraints for type inference

When doing conversion of universes, not only do we have =/<= constraints but also Lub constraints which correspond to unification of two levels which might not be necessary if unfolding is performed.

UWeak constraints come from irrelevant universes in cumulative polymorphism.

type t =

| QEq of Sorts.Quality.t * Sorts.Quality.t

| QLeq of Sorts.Quality.t * Sorts.Quality.t

| ULe of Sorts.t * Sorts.t

| UEq of Sorts.t * Sorts.t

| ULub of Univ.Level.t * Univ.Level.t

| UWeak of Univ.Level.t * Univ.Level.t

val is_trivial : t -> bool

val check_eq_level : UGraph.t -> Univ.Level.t -> Univ.Level.t -> bool

Wrapper around the UGraph function to handle Prop

module Set : sig ... end

type 'a constraint_function = 'a -> 'a -> Set.t -> Set.t

val enforce_eq_instances_univs : bool -> UVars.Instance.t constraint_function

val enforce_eq_qualities : Sorts.Quality.t array constraint_function

val compare_cumulative_instances : Conversion.conv_pb -> UVars.Variance.t array -> UVars.Instance.t constraint_function