Module UnivSubst

type 'a universe_map = 'a Univ.Level.Map.t

type universe_subst = Univ.Universe.t universe_map

type universe_subst_fn = Univ.Level.t -> Univ.Universe.t option

type universe_level_subst_fn = Univ.Level.t -> Univ.Level.t

val level_subst_of : universe_subst_fn -> universe_level_subst_fn

The resulting function must never be called on a level which would produce an algebraic.

val subst_univs_constraints : universe_subst_fn -> Univ.Constraints.t -> Univ.Constraints.t

val subst_instance : universe_level_subst_fn -> Univ.Instance.t -> Univ.Instance.t

type universe_opt_subst = Univ.Universe.t option universe_map

val normalize_univ_variables : universe_opt_subst -> universe_opt_subst * Univ.Level.Set.t * universe_subst

val normalize_univ_variable_opt_subst : universe_opt_subst -> Univ.Level.t -> Univ.Universe.t option

val normalize_universe_opt_subst : universe_opt_subst -> Univ.Universe.t -> Univ.Universe.t

val normalize_opt_subst : universe_opt_subst -> universe_opt_subst

val nf_binder_annot : (Sorts.relevance -> Sorts.relevance) -> 'a Context.binder_annot -> 'a Context.binder_annot

val nf_evars_and_universes_opt_subst : (Constr.existential -> Constr.constr option) -> (Univ.Level.t -> Univ.Level.t) -> (Sorts.t -> Sorts.t) -> (Sorts.relevance -> Sorts.relevance) -> Constr.constr -> Constr.constr

val subst_univs_universe : (Univ.Level.t -> Univ.Universe.t option) -> Univ.Universe.t -> Univ.Universe.t

val pr_universe_subst : (Univ.Level.t -> Pp.t) -> universe_subst -> Pp.t

val enforce_eq : Univ.Universe.t Univ.constraint_function

val enforce_leq : Univ.Universe.t Univ.constraint_function

val enforce_eq_sort : Sorts.t -> Sorts.t -> Univ.Constraints.t -> Univ.Constraints.t

val enforce_leq_sort : Sorts.t -> Sorts.t -> Univ.Constraints.t -> Univ.Constraints.t

val enforce_leq_alg_sort : Sorts.t -> Sorts.t -> UGraph.t -> Univ.Constraints.t * UGraph.t

Picks an arbitrary set of constraints sufficient to ensure u <= v.