Module Env.Path

type t

Paths are private, and owned by the functions below

val relative : t -> string -> t

relative path string build a path relative to an existing one

val exists : t -> bool

exists file checks if file exists

val to_string : t -> string

String representation