Module EConstr.ESorts

type t = Evd.esorts

Type of sorts up-to universe unification. Essentially a wrapper around Sorts.t so that normalization is ensured statically.

val make : Sorts.t -> t

Turn a sort into an up-to sort.

val kind : Evd.evar_map -> t -> Sorts.t

Returns the view into the current sort. Note that the kind of a variable may change if the unification state of the evar map changes.

val equal : Evd.evar_map -> t -> t -> bool

val is_small : Evd.evar_map -> t -> bool

val is_prop : Evd.evar_map -> t -> bool

val is_sprop : Evd.evar_map -> t -> bool

val is_set : Evd.evar_map -> t -> bool

val prop : t

val sprop : t

val set : t

val type1 : t

val super : Evd.evar_map -> t -> t

val relevance_of_sort : Evd.evar_map -> t -> Sorts.relevance

val family : Evd.evar_map -> t -> Sorts.family