Module Gramlib.Plexing

Lexing for Camlp5 grammars.

This module defines the Camlp5 lexer type to be used in extensible grammars (see module Grammar). It also provides some useful functions to create lexers.

Lexer type

module type S = sig ... end