Module Polynomial.Monomial

type t

A monomial is represented by a multiset of variables

val degree : t -> int

degree m is the sum of the degrees of each variable

val subset : t -> t -> bool

subset m1 m2 holds if the multi-set m1 is included in m2

val fold : (var -> int -> 'a -> 'a) -> t -> 'a -> 'a

fold f m acc folds f over the multiset m

val output : Stdlib.out_channel -> t -> unit

output o m outputs a textual representation