Module Reductionops.Stack

type 'a app_node

val pr_app_node : ('a -> Pp.t) -> 'a app_node -> Pp.t

type 'a case_stk = Constr.case_info * EConstr.EInstance.t * 'a array * 'a Constr.pcase_return * 'a Constr.pcase_invert * 'a Constr.pcase_branch array

type 'a member =

| App of 'a app_node

| Case of 'a case_stk

| Proj of Names.Projection.t

| Fix of ('a, 'a) Constr.pfixpoint * 'a t

| Primitive of CPrimitives.t * Names.Constant.t * EConstr.EInstance.t * 'a t * CPrimitives.args_red

and 'a t = 'a member list

val pr : ('a -> Pp.t) -> 'a t -> Pp.t

val empty : 'a t

val is_empty : 'a t -> bool

val decomp_node_last : 'a app_node -> 'a t -> 'a * 'a t

val compare_shape : 'a t -> 'a t -> bool

exception IncompatibleFold2

val fold2 : ('a -> EConstr.constr -> EConstr.constr -> 'a) -> 'a -> EConstr.constr t -> EConstr.constr t -> 'a

fold2 f x sk1 sk2 folds f on any pair of term in (sk1,sk2).

returns

the result and the lifts to apply on the terms

raises IncompatibleFold2

when sk1 and sk2 have incompatible shapes

val map : ('a -> 'a) -> 'a t -> 'a t

val append_app : 'a array -> 'a t -> 'a t

append_app args sk pushes array of arguments args on sk

val append_app_list : 'a list -> 'a t -> 'a t

append_app_list args sk pushes list of arguments args on sk

val strip_app : 'a t -> 'a t * 'a t

if strip_app sk = (sk1,sk2), then sk = sk1 @ sk2 with sk1 purely applicative and sk2 does not start with an argument

val strip_n_app : int -> 'a t -> ('a t * 'a * 'a t) option

returns

(the nth first elements, the (n+1)th element, the remaining stack) if there enough of those

val decomp : 'a t -> ('a * 'a t) option

decomp sk extracts the first argument of sk is there is some

val decomp_rev : 'a t -> ('a * 'a t) option

decomp sk extracts the first argument of reversed stack sk is there is some

val not_purely_applicative : 'a t -> bool

not_purely_applicative sk

val list_of_app_stack : EConstr.constr t -> EConstr.constr list option

list_of_app_stack sk either returns Some sk turned into a list of arguments if sk is purely applicative and None otherwise

val assign : 'a t -> int -> 'a -> 'a t

assign sk n a changes the nth argument of sk with a, counting from 0

raises an

anomaly if there is less that n arguments available

val args_size : 'a t -> int

args_size sk returns the number of arguments available at the head of sk

val tail : int -> 'a t -> 'a t

tail n sk drops the n first arguments of sk

raises [Invalid_argument]

if there are not enough arguments

val nth : 'a t -> int -> 'a

nth sk n returns the n-th argument of sk, counting from 0

raises [Not_found]

if there is no nth argument

val zip : Evd.evar_map -> (EConstr.constr * EConstr.constr t) -> EConstr.constr

zip sigma t sk