Module Structures.Instance

A canonical instance declares "canonical" conversion hints between the effective components of a structure and the projections of the structure

type t

val make : Environ.env -> Evd.evar_map -> Names.GlobRef.t -> t

Process a record instance, checkig it can be registered as canonical. The record type must be declared as a canonical Structure.t beforehand.

val register : warn:bool -> Environ.env -> Evd.evar_map -> t -> unit

Register an instance as canonical

val subst : Mod_subst.substitution -> t -> t

val repr : t -> Names.GlobRef.t