Module Structures.PrimitiveProjections

Some extra info for structures which are primitive records

val register : Names.Projection.Repr.t -> Names.Constant.t -> unit

Sets up the mapping from constants to primitive projections

val mem : Names.Constant.t -> bool

val find_opt : Names.Constant.t -> Names.Projection.Repr.t option