Module Structures.ValuePattern

A ValuePattern.t characterizes the form of a component of canonical instance and is used to query the data base of canonical instances

type t =

| Const_cs of Names.GlobRef.t

| Proj_cs of Names.Projection.Repr.t

| Prod_cs

| Sort_cs of Sorts.family

| Default_cs

val equal : Environ.env -> t -> t -> bool

val print : t -> Pp.t

val of_constr : Environ.env -> Constr.t -> t * int option * Constr.t list

Return the form of the component of a canonical structure