Module Vernacstate.System

System State

type t

The system state includes the summary and the libobject

val protect : ('a -> 'b) -> 'a -> 'b

protect f x runs f x and discards changes in the system state